+++ src/lib/components/events/ModalTicketsSell.svelte
@@ -1,26 +1,33 @@
 <script lang="ts">
   import { enhance } from "$app/forms";
   import { onMount } from "svelte";
+  import { TicketX, TicketCheck } from "lucide-svelte";
   import type { Event } from "$lib/types";
   import {
     calculatePrice,
     calculateTotalQuantity,
     calculateUbicationPrice,
   } from "$lib/utils/eventUtils";
+  import Loading from "../Loading.svelte";
+  import { fly, slide } from "svelte/transition";
   export let ticket: any;
   export let sellSystem: any;
   export let disclaimerEvent: any;
+  export let discountResponse: any = null;
+  export let discountCodeExist: any;
+
+  $: discountResponse, (isLoading = false);
+  $: discountResponse, handleChange();
+
+  let isLoading = false;
+  let haveDiscount = false;
+  let priceBeforeDiscount = 0;
 
   // ringside or general
   $: selectedTicketsType = "ringside_tickets";
   $: selectedTicketsQuantity = 1;
 
-  let selectedTicketsTotalPrice = 0;
-
-  $: formattedFinalPrice = new Intl.NumberFormat("es-CL", {
-    style: "currency",
-    currency: "CLP",
-  }).format(selectedTicketsTotalPrice);
+  $: selectedTicketsTotalPrice = 0;
 
   // options for select
   $: selectedTicketsTotalQuantity = calculateTotalQuantity(ticket);
@@ -29,24 +36,75 @@
     (_, i) => i + 1
   );
 
+  function applyDiscount(total, percentage) {
+    return total - total * (percentage / 100);
+  }
+
   const handleChange = () => {
     if (sellSystem === "ubication") {
       if (selectedTicketsType === "ringside_tickets") {
         selectedTicketsTotalPrice =
           selectedTicketsQuantity * ticket?.ringside_tickets?.price;
+        // Si existe un descuento se aplica
+        if (discountResponse?.success) {
+          priceBeforeDiscount = selectedTicketsTotalPrice;
+          selectedTicketsTotalPrice = applyDiscount(
+            selectedTicketsTotalPrice,
+            discountResponse?.percentage
+          );
+        }
       } else if (selectedTicketsType === "general_tickets") {
         selectedTicketsTotalPrice =
           selectedTicketsQuantity * ticket.general_tickets.price;
+        if (discountResponse?.success) {
+          priceBeforeDiscount = selectedTicketsTotalPrice;
+          selectedTicketsTotalPrice = applyDiscount(
+            selectedTicketsTotalPrice,
+            discountResponse?.percentage
+          );
+        }
       }
     } else {
       let obj = calculatePrice(selectedTicketsQuantity, ticket);
       selectedTicketsTotalPrice = obj.totalCost;
+
+      // Si existe un descuento se aplica
+      if (discountResponse?.success) {
+        priceBeforeDiscount = obj.totalCost;
+        selectedTicketsTotalPrice = applyDiscount(
+          selectedTicketsTotalPrice,
+          discountResponse?.percentage
+        );
+      }
     }
   };
 
   onMount(() => {
     handleChange();
   });
+
+  function toggleDiscount() {
+    haveDiscount = !haveDiscount;
+  }
+
+  // Reactive statement to identify ticket type with amount 0
+  $: ticketsWithZeroAmount = {
+    ringside: ticket?.ringside_tickets?.amount === 0,
+    general: ticket?.general_tickets?.amount === 0,
+  };
+
+  // Function to disable checkbox based on ticket type with amount 0
+  function isCheckboxDisabled(ticketType) {
+    return ticketsWithZeroAmount[ticketType];
+  }
+
+  $: {
+    if (ticketsWithZeroAmount.ringside && !ticketsWithZeroAmount.general) {
+      selectedTicketsType = 'general_tickets';
+    } else if (ticketsWithZeroAmount.general && !ticketsWithZeroAmount.ringside) {
+      selectedTicketsType = 'ringside_tickets';
+    }
+  }
 </script>
 
 <button
@@ -146,6 +204,7 @@
         </label>
         <select
           bind:value={selectedTicketsQuantity}
+          disabled={isLoading}
           name="tickets"
           id="tickets"
           required
@@ -171,6 +230,8 @@
                 name="ringside_tickets"
                 value="ringside_tickets"
                 class="radio radio-primary"
+                disabled={isCheckboxDisabled("ringside") || isLoading}
+                bind:group={selectedTicketsType}
                 checked={selectedTicketsType === "ringside_tickets"}
                 on:click={() => {
                   selectedTicketsType = "ringside_tickets";
@@ -185,6 +246,8 @@
               for="general_tickets"
             >
               <input
+                disabled={isCheckboxDisabled("general") || isLoading}
+                bind:group={selectedTicketsType}
                 type="radio"
                 id="general_tickets"
                 name="general_tickets"
@@ -212,11 +275,107 @@
         id="ticketsType"
         bind:value={selectedTicketsType}
       />
+
+      {#if discountCodeExist && !discountResponse?.success}
+        <div class="form-control" out:slide>
+          <label class="cursor-pointer label flex justify-start gap-2">
+            <input
+              disabled={isLoading}
+              type="checkbox"
+              class="checkbox checkbox-warning"
+              on:click={toggleDiscount}
+            />
+            <span class="label-text">¿Tienes un código de descuento?</span>
+          </label>
+        </div>
+        {#if isLoading}
+          <div class="flex gap-2" in:fly={{ y: 20 }} out:slide>
+            <Loading /> Validando código...
+          </div>
+        {:else if haveDiscount}
+          <form
+            action="?/validate"
+            method="POST"
+            in:fly={{ y: 20 }}
+            out:slide
+            use:enhance
+            on:submit={() => {
+              isLoading = true;
+              handleChange();
+            }}
+          >
+            <div class="form-control w-full">
+              <label class="label" for="discount">
+                <span class="label-text">¡Amig@ dale!, agrega tú código</span>
+              </label>
+              <div class="flex gap-2">
+                <input
+                  type="text"
+                  name="discount"
+                  id="discount"
+                  class="input input-bordered w-full"
+                />
+                <button class="btn btn-outline btn-primary hover:text-black"
+                  >Validar</button
+                >
+              </div>
+              {#if !discountResponse?.success && discountResponse?.success !== undefined}
+                <p
+                  class="text-error mt-2 flex gap-2
+                "
+                >
+                  <TicketX />
+                  {discountResponse?.message}
+                </p>
+              {/if}
+            </div>
+          </form>
+        {/if}
+      {/if}
+
+      {#if discountResponse?.success}
+        <p class="mt-2 text-success flex gap-2" in:fly={{ y: 20 }}>
+          <TicketCheck />
+          {discountResponse?.message}
+        </p>
+      {/if}
+
+      <!-- Envio el codigo al servidor -->
+      {#if discountResponse?.success}
+        <input
+          type="hidden"
+          name="discountCode"
+          id="discountCode"
+          value={discountResponse?.code}
+        />
+      {/if}
+
+      <div class="divider divider-neutral text-white">Total</div>
+
+      <p class="text-center">
+        <span class="line-through text-error" in:fly={{ y: 20 }}
+          >{priceBeforeDiscount
+            ? new Intl.NumberFormat("es-CL", {
+                style: "currency",
+                currency: "CLP",
+              }).format(priceBeforeDiscount)
+            : ""}</span
+        >
+        <span class="text-2xl text-primary"
+          >{new Intl.NumberFormat("es-CL", {
+            style: "currency",
+            currency: "CLP",
+          }).format(selectedTicketsTotalPrice)}</span
+        >
+      </p>
+
       <button
         type="submit"
+        disabled={isLoading}
         class="flex grow w-full items-center rounded-none btn btn-primary cursor-pointer text-black no-underline col-span-2"
-        >{formattedFinalPrice} - Comprar</button
       >
+        Comprar
+      </button>
     </form>
   </div>
   <form method="dialog" class="modal-backdrop">