+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fix
+
+- Check radiobox when the other option is disabled
+
+### Add
+
+- Discount codes
+
 ## [0.0.32] - 2024-06-24
 
 ### Add